requirements.txt
openai
requests
python-dotenv
duckduckgo-search
pandas
colorama